Allow ECMA arrays to encode Bools and Numbers if prefered
Limited by the Swift type system here as you cannot declare that the
value in a dictionary is Codable (Encodable) and it just deals with it.
I can't figure out a better way of dealing with the mixed types that
can occur inside an ECMA array other than something like this. Nasty.